q2_multimodal_qa_agent/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
requests
python-dotenv
google-generativeai
